apps/text-generator-service/pyproject.toml
[tool.black]
line-length = 125
skip-string-normalization = true

[tool.poetry]
name = "text-generator-service"
version = "0.1.0"
description = ""
authors = ["Kevin Yu <kevin.yu.404@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10, <3.12"
grpcio = "^1.51.1"
keras = "^2.11.0"
tensorflow = "^2.11.0"
protobuf = ">=3.9.2,<3.20"
grpcio-tools = "1.48.2"


[tool.poetry.group.dev.dependencies]
radon = "^5.1.0"
flake8 = "^6.0.0"
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
